third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-script-preprocessor-expected.txt

CONSOLE MESSAGE: line 2: preprocessed script1.js.js
CONSOLE MESSAGE: line 86: preprocessed debugger-script-preprocessor_html.js
CONSOLE MESSAGE: line 2: eval works
CONSOLE MESSAGE: line 2: verify href: debugger-script-preprocessor.html
CONSOLE ERROR: line 18: Uncaught Error: fake internal preprocessor error
CONSOLE MESSAGE: line 94: preprocessed debugger-script-preprocessor_html_onload.js
CONSOLE ERROR: line 1: Uncaught Error: failingPreprocessor throws
CONSOLE ERROR: The preprocessor must compile to a function.
CONSOLE ERROR: The preprocessor must compile to a function.
CONSOLE ERROR: line 2: Uncaught SyntaxError: Unexpected identifier
CONSOLE ERROR: The preprocessor must compile to a function.
Tests script preprocessor (ability to preprocess all scripts upon reload).

Page reloaded.
sourceURLs[0]=inspector-test.js
sourceURLs[1]=debugger-test.js
sourceURLs[2]=script1.js.js
sourceURLs[3]=debugger-script-preprocessor_html.js
sourceURLs[4]=debugger-script-preprocessor_html_onload.js
sourceURLs[5]=dynamicScript1
sourceURLs[6]=dynamicScript2
Page reloaded.
Page reloaded.
Page reloaded.
Page reloaded.
